Completed exceptions, made minor adjustments to controller/service/mo…

+++ src/main/java/com/bankapi/bankofmikaila/service/DepositService.java
@@ -1,6 +1,12 @@
 package com.bankapi.bankofmikaila.service;
 
+import com.bankapi.bankofmikaila.exceptions.DepositByAccountNotFound;
+import com.bankapi.bankofmikaila.exceptions.DepositByIdNotFound;
+import com.bankapi.bankofmikaila.exceptions.WithdrawalByIdNotFound;
+import com.bankapi.bankofmikaila.exceptions.WithdrawlsByAccountNotFound;
+import com.bankapi.bankofmikaila.model.Account;
 import com.bankapi.bankofmikaila.model.Deposit;
+import com.bankapi.bankofmikaila.model.Withdrawl;
 import com.bankapi.bankofmikaila.repository.AccountRepository;
 import com.bankapi.bankofmikaila.repository.DepositRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,13 +23,32 @@ public class DepositService {
     @Autowired
     private AccountRepository accountRepository;
 
-    public Iterable<Deposit> getAllDeposits(){
 
-        return depositRepository.findAll();
+    protected void verifyAccount(Long accountId) throws DepositByAccountNotFound {
+
+        accountRepository.findById(accountId).orElseThrow(()->
+                new DepositByAccountNotFound("Account not found."));
+
+    }
+
+    protected void verifyDeposit(Long depositId) throws DepositByIdNotFound {
+
+        depositRepository.findById(depositId).orElseThrow(()->
+                new DepositByIdNotFound("Error fetching deposit with id."));
+    }
+
+
+    public Iterable<Deposit> getAllDeposits(Long accountId){
+
+        verifyAccount(accountId);
+
+        return depositRepository.findDepositsByAccountId(accountId);
 
     }
 
     public Optional<Deposit> getDeposit(Long depositId){
+
+        verifyDeposit(depositId);
 
         return depositRepository.findById(depositId);
 
@@ -32,9 +57,9 @@ public class DepositService {
     public void createDeposit(Deposit deposit, Long accountId) {
 
         //uses lambda expression to check if account exists (by id), if not throw exception
-        accountRepository.findById(accountId).orElseThrow(()->
-                new EntityNotFoundException("Account with Id: " + accountId + " not found."));
-
+        Account account = accountRepository.findById(accountId).orElseThrow(()->
+                new DepositByAccountNotFound("Error creating deposit: account not found."));
+        deposit.setAccount(account);
         //Save deposit once null check passes
         depositRepository.save(deposit);
 
@@ -60,14 +85,20 @@ public class DepositService {
 
             depositRepository.save(existingDeposit);
         } else {
-            throw new EntityNotFoundException("Deposit with Id: " + depositId + " not found.");
+            throw new DepositByIdNotFound("Deposit ID does not exist.");
         }
 
     }
 
     public void deleteDeposit(Long depositId){
 
-        depositRepository.deleteById(depositId);
+        Optional<Deposit> existingDepositOptional = depositRepository.findById(depositId);
+
+        if(existingDepositOptional.isPresent()) {
+            depositRepository.deleteById(depositId);
+        } else {
+            throw new DepositByIdNotFound("This id does not exist in deposits");
+        }
 
     }
